+++ autosub/core.py
@@ -8,7 +8,6 @@ from __future__ import absolute_import, print_function, unicode_literals
 
 # Import built-in modules
 import os
-import wave
 import multiprocessing
 
 # Import third-party modules
@@ -18,7 +17,7 @@ import auditok
 
 # Any changes to the path and your own modules
 from autosub import speech_trans_api
-from autosub import formatters
+from autosub import sub_utils
 from autosub import constants
 from autosub import ffmpeg_utils
 
@@ -28,7 +27,7 @@ def auditok_gen_speech_regions(  # pylint: disable=too-many-arguments
         energy_threshold=constants.DEFAULT_ENERGY_THRESHOLD,
         min_region_size=constants.MIN_REGION_SIZE,
         max_region_size=constants.MAX_REGION_SIZE,
-        max_continuous_silence=constants.MAX_CONTINUOUS_SILENCE,
+        max_continuous_silence=constants.DEFAULT_CONTINUOUS_SILENCE,
         mode=auditok.StreamTokenizer.STRICT_MIN_LENGTH
 ):
     """
@@ -62,133 +61,108 @@ def auditok_gen_speech_regions(  # pylint: disable=too-many-arguments
     return regions
 
 
-def sub_gen_speech_regions(
-        source_file,
-        sub_file,
-        ext_max_size_ms=constants.MAX_EXT_REGION_SIZE * 1000
-):
-    """
-    Given an input audio/video file and subtitles file, generate proper speech regions.
-    """
-    regions = []
-    audio_wav = ffmpeg_utils.source_to_audio(source_file)
-    reader = wave.open(audio_wav)
-    audio_file_length = int(float(reader.getnframes()) / float(reader.getframerate())) * 1000
-    reader.close()
-
-    ext_regions = pysubs2.SSAFile.load(sub_file)
-
-    for event in ext_regions.events:
-        if not event.is_comment:
-            # not a comment region
-            if event.duration <= ext_max_size_ms:
-                regions.append((event.start,
-                                event.start + event.duration))
-            else:
-                # split too long regions
-                elapsed_time = event.duration
-                start_time = event.start
-                if elapsed_time > audio_file_length:
-                    # even longer than the source file length
-                    elapsed_time = audio_file_length
-                while elapsed_time > ext_max_size_ms:
-                    # longer than the max size limit
-                    regions.append((start_time,
-                                    start_time + ext_max_size_ms))
-                    elapsed_time = elapsed_time - ext_max_size_ms
-                    start_time = start_time + ext_max_size_ms
-                regions.append((start_time,
-                                start_time + elapsed_time))
-
-    os.remove(audio_wav)
-    return regions
-
-
-def api_gen_text(  # pylint: disable=too-many-locals,too-many-arguments,too-many-branches,too-many-statements
+def speech_to_text(  # pylint: disable=too-many-locals,too-many-arguments,too-many-branches,too-many-statements
         source_file,
         api_url,
         regions,
-        api_key,
         concurrency=constants.DEFAULT_CONCURRENCY,
         src_language=constants.DEFAULT_SRC_LANGUAGE,
-        dst_language=constants.DEFAULT_DST_LANGUAGE
+        min_confidence=0.0
 ):
     """
-    Given an input audio/video file, generate subtitles in the specified language and format.
+    Given an input audio/video file, generate text_list from speech-to-text api.
     """
+
+    if not regions:
+        return None
+
     audio_rate = 44100
     audio_flac = ffmpeg_utils.source_to_audio(source_file, rate=audio_rate, file_ext='.flac')
     pool = multiprocessing.Pool(concurrency)
     converter = ffmpeg_utils.SplitIntoFLACPiece(source_path=audio_flac)
 
     recognizer = speech_trans_api.GoogleSpeechToTextV2(
-        language=src_language,
-        rate=audio_rate,
         api_url=api_url,
+        min_confidence=min_confidence,
+        lang_code=src_language,
+        rate=audio_rate,
         api_key=constants.GOOGLE_SPEECH_V2_API_KEY)
 
-    transcripts = []
-    if regions:
-        widgets = ["Converting speech regions to FLAC files: ",
+    text_list = []
+    widgets = ["Converting speech regions to FLAC files: ",
+               progressbar.Percentage(), ' ',
+               progressbar.Bar(), ' ',
+               progressbar.ETA()]
+    pbar = progressbar.ProgressBar(widgets=widgets, maxval=len(regions)).start()
+    try:
+        flac_regions = []
+        for i, flac_region in enumerate(pool.imap(converter, regions)):
+            flac_regions.append(flac_region)
+            pbar.update(i)
+        pbar.finish()
+
+        widgets = ["Performing speech recognition: ",
                    progressbar.Percentage(), ' ',
                    progressbar.Bar(), ' ',
                    progressbar.ETA()]
         pbar = progressbar.ProgressBar(widgets=widgets, maxval=len(regions)).start()
-        try:
-            extracted_regions = []
-            for i, extracted_region in enumerate(pool.imap(converter, regions)):
-                extracted_regions.append(extracted_region)
-                pbar.update(i)
-            pbar.finish()
 
-            widgets = ["Performing speech recognition: ",
-                       progressbar.Percentage(), ' ',
-                       progressbar.Bar(), ' ',
-                       progressbar.ETA()]
-            pbar = progressbar.ProgressBar(widgets=widgets, maxval=len(regions)).start()
+        for i, transcript in enumerate(pool.imap(recognizer, flac_regions)):
+            text_list.append(transcript)
+            pbar.update(i)
+        pbar.finish()
 
-            for i, transcript in enumerate(pool.imap(recognizer, extracted_regions)):
-                transcripts.append(transcript)
-                pbar.update(i)
-            pbar.finish()
-
-            if src_language.split("-")[0] != dst_language.split("-")[0]:
-                if api_key:
-                    google_translate_api_key = api_key
-                    translator = \
-                        speech_trans_api.GoogleTranslatorV2(dst_language,
-                                                            google_translate_api_key,
-                                                            dst=dst_language,
-                                                            src=src_language)
-                    prompt = "Translating from {0} to {1}: ".format(src_language, dst_language)
-                    widgets = [prompt, progressbar.Percentage(), ' ',
-                               progressbar.Bar(), ' ',
-                               progressbar.ETA()]
-                    pbar = progressbar.ProgressBar(widgets=widgets, maxval=len(regions)).start()
-                    translated_transcripts = []
-                    for i, transcript in enumerate(pool.imap(translator, transcripts)):
-                        translated_transcripts.append(transcript)
-                        pbar.update(i)
-                    pbar.finish()
-                    transcripts = translated_transcripts
-                else:
-                    print(
-                        "Error: Subtitle translation requires specified Google Translate API key. "
-                        "See --help for further information."
-                    )
-                    return 1
-
-        except KeyboardInterrupt:
-            pbar.finish()
-            pool.terminate()
-            pool.join()
-            print("Cancelling transcription.")
-            return 1
+    except KeyboardInterrupt:
+        pbar.finish()
+        pool.terminate()
+        pool.join()
+        print("Cancelling transcription.")
+        return 1
 
     os.remove(audio_flac)
-    timed_subtitles = [(r, t) for r, t in zip(regions, transcripts) if t]
+    return text_list
 
-    return timed_subtitles
+
+def text_translation(
+        text_list,
+        api_key,
+        concurrency=constants.DEFAULT_CONCURRENCY,
+        src_language=constants.DEFAULT_SRC_LANGUAGE,
+        dst_language=constants.DEFAULT_DST_LANGUAGE
+):
+    """
+    Given a text list, generate translated text list from translation api.
+    """
+
+    if not text_list:
+        return None
+
+    pool = multiprocessing.Pool(concurrency)
+    google_translate_api_key = api_key
+    translator = \
+        speech_trans_api.GoogleTranslatorV2(dst_language,
+                                            google_translate_api_key,
+                                            dst=dst_language,
+                                            src=src_language)
+    prompt = "Translating from {0} to {1}: ".format(src_language, dst_language)
+    widgets = [prompt, progressbar.Percentage(), ' ',
+               progressbar.Bar(), ' ',
+               progressbar.ETA()]
+    pbar = progressbar.ProgressBar(widgets=widgets, maxval=len(text_list)).start()
+    try:
+        translated_text = []
+        for i, transcript in enumerate(pool.imap(translator, text_list)):
+            translated_text.append(transcript)
+            pbar.update(i)
+        pbar.finish()
+    except KeyboardInterrupt:
+        pbar.finish()
+        pool.terminate()
+        pool.join()
+        print("Cancelling transcription.")
+        return 1
+
+    return translated_text
 
 
 def list_to_sub_str(  # pylint: disable=too-many-arguments
@@ -203,7 +177,7 @@ def list_to_sub_str(  # pylint: disable=too-many-arguments
 
     if subtitles_file_format == 'srt' \
             or subtitles_file_format == 'tmp':
-        formatted_subtitles = formatters.pysubs2_formatter(
+        formatted_subtitles = sub_utils.pysubs2_formatter(
             subtitles=timed_subtitles,
             sub_format=subtitles_file_format)
 
@@ -214,26 +188,26 @@ def list_to_sub_str(  # pylint: disable=too-many-arguments
             ass_styles = ass_file.styles
         else:
             ass_styles = None
-        formatted_subtitles = formatters.pysubs2_formatter(
+        formatted_subtitles = sub_utils.pysubs2_formatter(
             subtitles=timed_subtitles,
             sub_format=subtitles_file_format,
             ass_styles=ass_styles)
 
     elif subtitles_file_format == 'vtt':
-        formatted_subtitles = formatters.vtt_formatter(
+        formatted_subtitles = sub_utils.vtt_formatter(
             subtitles=timed_subtitles)
 
     elif subtitles_file_format == 'json':
-        formatted_subtitles = formatters.json_formatter(
+        formatted_subtitles = sub_utils.json_formatter(
             subtitles=timed_subtitles)
 
     elif subtitles_file_format == 'txt':
-        formatted_subtitles = formatters.txt_formatter(
+        formatted_subtitles = sub_utils.txt_formatter(
             subtitles=timed_subtitles)
 
     elif subtitles_file_format == 'sub':
         subtitles_file_format = 'microdvd'
-        formatted_subtitles = formatters.pysubs2_formatter(
+        formatted_subtitles = sub_utils.pysubs2_formatter(
             subtitles=timed_subtitles,
             sub_format=subtitles_file_format,
             fps=fps)
@@ -243,7 +217,7 @@ def list_to_sub_str(  # pylint: disable=too-many-arguments
         subtitles_file_format = 'sub'
 
     elif subtitles_file_format == 'mpl2':
-        formatted_subtitles = formatters.pysubs2_formatter(
+        formatted_subtitles = sub_utils.pysubs2_formatter(
             subtitles=timed_subtitles,
             sub_format=subtitles_file_format)
         subtitles_file_format = 'mpl2.txt'
@@ -253,11 +227,11 @@ def list_to_sub_str(  # pylint: disable=too-many-arguments
         print("Format \"{fmt}\" not supported. \
         Using \"{default_fmt}\" instead.".format(fmt=subtitles_file_format,
                                                  default_fmt=constants.DEFAULT_SUBTITLES_FORMAT))
-        formatted_subtitles = formatters.pysubs2_formatter(
+        formatted_subtitles = sub_utils.pysubs2_formatter(
             subtitles=timed_subtitles,
             sub_format=constants.DEFAULT_SUBTITLES_FORMAT)
 
-    return formatted_subtitles
+    return formatted_subtitles, subtitles_file_format
 
 
 def times_to_sub_str(  # pylint: disable=too-many-arguments
@@ -272,7 +246,7 @@ def times_to_sub_str(  # pylint: disable=too-many-arguments
 
     if subtitles_file_format == 'srt' \
             or subtitles_file_format == 'tmp':
-        formatted_subtitles = formatters.pysubs2_times_formatter(
+        formatted_subtitles = sub_utils.pysubs2_times_formatter(
             times=times,
             sub_format=subtitles_file_format)
 
@@ -283,22 +257,22 @@ def times_to_sub_str(  # pylint: disable=too-many-arguments
             ass_styles = ass_file.styles
         else:
             ass_styles = None
-        formatted_subtitles = formatters.pysubs2_times_formatter(
+        formatted_subtitles = sub_utils.pysubs2_times_formatter(
             times=times,
             sub_format=subtitles_file_format,
             ass_styles=ass_styles)
 
     elif subtitles_file_format == 'vtt':
-        formatted_subtitles = formatters.vtt_times_formatter(
+        formatted_subtitles = sub_utils.vtt_times_formatter(
             times=times)
 
     elif subtitles_file_format == 'json':
-        formatted_subtitles = formatters.json_times_formatter(
+        formatted_subtitles = sub_utils.json_times_formatter(
             times=times)
 
     elif subtitles_file_format == 'sub':
         subtitles_file_format = 'microdvd'
-        formatted_subtitles = formatters.pysubs2_times_formatter(
+        formatted_subtitles = sub_utils.pysubs2_times_formatter(
             times=times,
             sub_format=subtitles_file_format,
             fps=fps)
@@ -308,7 +282,7 @@ def times_to_sub_str(  # pylint: disable=too-many-arguments
         subtitles_file_format = 'sub'
 
     elif subtitles_file_format == 'mpl2':
-        formatted_subtitles = formatters.pysubs2_times_formatter(
+        formatted_subtitles = sub_utils.pysubs2_times_formatter(
             times=times,
             sub_format=subtitles_file_format)
         subtitles_file_format = 'mpl2.txt'
@@ -318,7 +292,7 @@ def times_to_sub_str(  # pylint: disable=too-many-arguments
         print("Format \"{fmt}\" not supported. \
         Using \"{default_fmt}\" instead.".format(fmt=subtitles_file_format,
                                                  default_fmt=constants.DEFAULT_SUBTITLES_FORMAT))
-        formatted_subtitles = formatters.pysubs2_times_formatter(
+        formatted_subtitles = sub_utils.pysubs2_times_formatter(
             times=times,
             sub_format=constants.DEFAULT_SUBTITLES_FORMAT)